backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
anthropic==0.40.0
PyMuPDF==1.24.11
Pillow==11.0.0
numpy==2.1.3
pydantic==2.9.2
python-dotenv==1.0.1
